refactor: make CombatantLogic Foundry-pure and fold in combatant-logic.ts
Now that CombatantLogic lives under combatant/logic/, treat it as a pure-logic
module like the actor/item logic classes: no direct Foundry document reads.

De-spill: the relational/scene-coupled reads that touched the combatant
document directly now go through the data port or a FoundryHelpers shim.
- New derived facts on the combatant data port (SohlCombatantDataModel):
  groupId, isDefeated, statuses, isHidden. CombatantLogic reads these via
  this.data / peer cl.data instead of combatant._source.group, actor.statuses,
  isDefeated, token.hidden.
- New shims fvttCombatantLogics (peer logics) and fvttPromptMoveCombatantToGroup
  replace the combatant.combat.combatants walk and the direct moveToGroup call.
- Card actorUuid uses this.actorLogic.uuid instead of combatant.actor.uuid.
- combatant/logic/** is added to the ESLint Foundry-free zone and the purity
  smoke test, so this is now enforced.

Fold: combatant-logic.ts's pure helpers (areCombatantsEnemies, isThreatening,
computeMove, chooseInitialDisplayedMedium, THREAT_NEGATING_STATUSES) move into
CombatantLogic.ts as module-level exports; the file is deleted and importers
(SohlCombatant, two tests) updated.

Also completes the trailing import-path updates from the prior combatant
folder move (BeingLogic, automated-combat, SohlCombat, sohl.ts).
