+++ src/document/actor/logic/BeingLogic.ts
@@ -41,7 +41,7 @@ import {
 } from "@src/document/actor/logic/token-helpers";
 import { getActiveScene, getActiveCombat } from "@src/core/FoundryHelpers";
 import type { SohlTokenDocument } from "@src/document/token/SohlTokenDocument";
-import type { SohlCombatant } from "@src/document/combatant/SohlCombatant";
+import type { SohlCombatant } from "@src/document/combatant/foundry/SohlCombatant";
 import { readBaseMove } from "@src/domain/movement/move-helpers";
 import {
     ACTION_SUBTYPE,
